+++ Dochazka.Tests/UnitTests/PaginatedListTests.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Dochazka.HelperClasses;
+using Dochazka.Models;
 using Xunit;
 
 namespace Dochazka.Tests.UnitTests
-{    
+{
     public class PaginatedListTests
     {
         public List<int> testList;
@@ -26,7 +26,7 @@ namespace Dochazka.Tests.UnitTests
             var expectedResult = new List<int> { 1, 2, 3, 4 };
 
             // 2. Act
-            var actualResult = PaginatedList<int>.Create(testList.AsQueryable(), pageIndex ?? 1, pageSize);
+            var actualResult = PaginatedListViewModel<int>.Create(testList.AsQueryable(), pageIndex ?? 1, pageSize);
 
             // 3. Assert
             Assert.Equal(expectedResult.Count, actualResult.Count);
@@ -50,7 +50,7 @@ namespace Dochazka.Tests.UnitTests
             var expectedResult = new List<int> { 9, 0 };
 
             // 2. Act
-            var actualResult = PaginatedList<int>.Create(testList.AsQueryable(), pageNumber ?? 1, pageSize);
+            var actualResult = PaginatedListViewModel<int>.Create(testList.AsQueryable(), pageNumber ?? 1, pageSize);
 
             // 3. Assert
             Assert.Equal(expectedResult.Count, actualResult.Count);
