+++ Dochazka/HelperClasses/CSVResult.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Text;
-using System.Data;
+﻿using System.Data;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Http;
 using System.Threading.Tasks;
