+++ Dochazka/Views/AttendanceRecords/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model PaginatedList<Dochazka.Models.AttendanceRecordModel>
+﻿@model PaginatedListViewModel<Dochazka.Models.AttendanceRecordModel>
 
 @{
     bool managerControls = false;
