+++ Dochazka/Views/Teams/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model PaginatedList<Dochazka.Models.TeamModel>
+﻿@model PaginatedListViewModel<Dochazka.Models.TeamModel>
 
 @{
     ViewData["Title"] = "Index";
@@ -55,24 +55,21 @@
     </table>
 }
 
-
-
-
 @{
     var prevDisabled = !Model.HasPreviousPage ? "disabled" : "";
     var nextDisabled = !Model.HasNextPage ? "disabled" : "";
 }
 
 <a asp-action="Index"
-   asp-route-sortOrder="@ViewData["CurrentSort"]"
    asp-route-pageNumber="@(Model.PageIndex - 1)"
+   asp-route-sortOrder="@ViewData["CurrentSort"]"   
    asp-route-currentFilter="@ViewData["CurrentFilter"]"
    class="btn btn-default @prevDisabled">
     Previous
 </a>
 <a asp-action="Index"
-   asp-route-sortOrder="@ViewData["CurrentSort"]"
    asp-route-pageNumber="@(Model.PageIndex + 1)"
+   asp-route-sortOrder="@ViewData["CurrentSort"]"   
    asp-route-currentFilter="@ViewData["CurrentFilter"]"
    class="btn btn-default @nextDisabled">
     Next
