+++ Dochazka/Views/UserRoles/Delete.cshtml
@@ -48,6 +48,7 @@
             <input type="submit" value="Delete" class="btn btn-danger" />
             <text> | </text>
         }
+
         <a asp-action="Index">Back to List</a>
     </form>
 </div>
